requirements.txt
flask
flask_login
flask_migrate
flask_sqlalchemy
python-dotenv
psycopg2-binary
requests